requirements.txt
django==4.0.1
django-simpleui==2022.1
django-ckeditor==6.2.0
gunicorn==20.1.0
pymysql==1.0.2
requests==2.27.1
loguru==0.5.3
arrow==1.2.1
weixin-python==0.5.7
wechatpayv3==1.2.13
django-cors-headers==3.11.0
pymysql==1.0.2
emoji
pillow
